requirements.txt
Flask==2.0.1
pymongo==4.6.3
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.2
Werkzeug==2.3.8